Replace the internal `BitPtr` representation
This commit replaces the element/tail tuple in `BitPtr` with a single
bit-count value. The previous design was used in order to provide fast
access to the tail index. However, the element/tail tuple created a more
complex value space and process of modifying a `BitPtr` range.

The head and length pattern still allows fast, linear computation of the
tail index, and makes the value space and modification process much more
straightforward.