Encode type widths in the `BitIdx` type, to reduce runtime assertions
This commit makes `BitIdx` generic over a `BitStore` parameter, and
restricts public constructors to only use assertive functions. Once a
`BitIdx` is created, the crate can safely manipulate it without any
assertions, as the crate logic respects boundaries throughout, and does
not need to double-check this work.

This change had the side effect of requiring that `BitIdx` no longer be
used to carry tail indices, which must now be split into a distinct
type. The `TailIdx` type propagated through tail analysis, and has the
minimum API needed to make the crate operable.