`BitSlice::{s,g}et_unchecked` no longer produce naïve references
As it is UB to create a standard reference to a primitive that may be
aliased, even if it is immediately reinterpreted as a reference to an
aliasable type, the last sites where aliased bare references may be
produced are altered to instead produce references to the build-specific
aliasable type. This should remove the last concurrency-related fault
conditions.