Restore the `__bits_from_slice!` macro and start a test suite
The `__bits_from_slice!` macro ensures that the literal tokens used to
accelerate known orderings are correctly mapped to `bitvec` types, and
not to any other names that may be in the caller's scope.

Because the fundamental types' `from_*e_bytes` constructors are not
`const fn`, their presence in literal token sequences did not permit the
compiler to elevate the literal to the `'static` lifetime. This meant
that using `bits!` to return a `&BitSlice` reference over the arrays
generated by the macro failed, as the array literals went out of scope
after the `&BitSlice` constructor call exited rather than being hoisted
for the duration of the reference's existence.

In order to circumvent this, the `from_*e_bytes` constructors are
explicitly written out as `const fn` items, and a new helper macro is
written in order to dispatch token chunks to the correct constructor.
This enables the `bits!` macros to create `static DATA` bindings for the
constructed array literals, and thus make the macro return
`&'static BitSlice`.

In the future, the syntax may be extended to permit mutable buffers.