Tighten restrictions on memory aliasing
This removes the possibility of constructing `BitSlice`s over memory
that may be externally aliased. This is necessary because the
`domain` module is able to deälias memory, but is only permitted to
do so when `bitvec` has sole and complete knowledge of the aliasing
status of memory regions under its control.