Fix miri failures by preserving provenance in `BitSpan::new_unchecked`
Fix the miri failures in doctests, see issue #135. The issue is that miri doesn't guess
correct provenance in the int-to-ptr cast in `BitSpan::new_unchecked`, as was found by
@tavianator [here](https://github.com/rust-lang/miri/issues/1866#issuecomment-985770125).

The solution is to preserve provenance and was proposed by  @tavianator
[here](https://github.com/rust-lang/miri/issues/1866#issuecomment-986068198).
With this change the entire test suite passes under miri.