Fix GitHub Actions workflow
The current version of the action `peaceiris/actions-mdbook` only
supports Ubuntu 22.04 or 20.04, but not Ubuntu 18.04 anymore. So
an update of the runner from Ubuntu 18.04 to Ubuntu 22.04 is
probably the best way to fix this.

Furthermore, the action `actions/checkout` is updated to get rid
of a deprecation warning related to an old Node.js version.